Controllers listos

+++ src/main/java/com/seguimiento/model/Categoria.java
@@ -8,11 +8,11 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
     private String nombre;
 
     public Categoria() {
-        //
+        
     }
 
     public Categoria(String nombre) {
@@ -20,10 +20,10 @@ public class Categoria {
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
